devtools/client/webconsole/new-console-output/test/components/test_evaluation-result.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for EvaluationResult component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for EvaluationResult component</p>

<script type="text/javascript;version=1.8">
window.onload = Task.async(function* () {
  const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
  const {
    EvaluationResult,
    getPreviewComponent
  } = require("devtools/client/webconsole/new-console-output/components/message-types/evaluation-result");

  yield testFullRender();
  yield testGetPreviewComponent();

  SimpleTest.finish()

  /**
   * Test that passing in a message correctly wires up all the children.
   *
   * The different combinations of children are tested in separate per-component
   * tests. This test just ensures that this component pipes data to its children.
   */
  function testFullRender() {
    const testValue = testCommands.get("new Date()");
    const packet = yield getPacket(testValue.command, testValue.commandType);
    const message = prepareMessage(packet);
    const props = {
      message
    };
    const rendered = renderComponent(EvaluationResult, props);

    ok(rendered.textContent.includes(testValue.expectedText),
      "EvaluationResult pipes data to its children as expected");
}

  /**
   * Test that getPreviewComponent() returns correct component for each object type.
   */
  function testGetPreviewComponent() {
    const testValues = [
      {
        commandObj: testCommands.get("new Date()"),
        expectedComponent: require("devtools/client/webconsole/new-console-output/components/message-types/date-preview").DatePreview
      }
    ];

    for (let testValue of testValues) {
      const { commandObj, expectedComponent } = testValue;
      const packet = yield getPacket(commandObj.command, commandObj.commandType);
      const message = prepareMessage(packet);
      const rendered = shallowRenderComponent(EvaluationResult, {message});
      is(rendered.type, expectedComponent,
        `EvaluationResult nests ${expectedComponent} based on command: ${commandObj.command}`);
    }
  }
});

</script>
</body>
</html>
